requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.10.4

# PDF processing
pypdf==4.3.1
pdfplumber==0.11.4
camelot-py[cv]==0.11.0

# Data handling
pandas==2.2.3
pyarrow==18.1.0
numpy==1.26.4

# Vector store
qdrant-client==1.12.1

# LangChain ecosystem
langchain>=1.2.0
langchain-community>=1.0.0
langchain-text-splitters>=1.0.0
langchain-openai>=1.0.0
langchain-google-genai>=1.0.0
langchain-core>=1.2.0

# Embeddings and ML
sentence-transformers==3.3.1
transformers==4.47.1

# Web interface
streamlit==1.41.1

# Optional: Jupyter notebooks
# jupyter==1.1.1
# ipykernel==6.29.5
